+++ actioncable/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Rails 8.0.0.rc1 (October 19, 2024) ##
+
+*   No changes.
+
+
 ## Rails 8.0.0.beta1 (September 26, 2024) ##
 
 *   Add an `identifier` to the event payload for the ActiveSupport::Notification `transmit_subscription_confirmation.action_cable` and `transmit_subscription_rejection.action_cable`.